drm/tidss: add get_irq() to dispc_ops
tidss supports devices which have multiple instances
of IRQs and any one can be selected for driving the
device.

For example, there are 4 mirrored interrupt regions
in J721E, all of which are identical and can be
used for interrupt handling.

Adding get_irq() function to dispc_ops so that
the device specific code can decide which IRQ to use

Signed-off-by: Subhajit Paul <subhajit_paul@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>