dmaengine: ti: k3-udma: Support for channel performance levels
Instead of the deprecated bool the binding documentation tells that the
performance level of a channel to be used can be selected via the
ti,channel-tpl

The available performance levels can differ between UDMAP implementations
and the number of channels also for the levels can also.

Use match data to provide a simple table for the tpl of the given UDMAP
and use that in the driver to safely select the level requested by the
client.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>