HACK: drm/omap: increase DSS5 max tv pclk to 192MHz
DSS5's maximum tv pclk rate (i.e. HDMI) is set to 186MHz, which comes
from the TRM (DPLL_HDMI_CLK1 frequency must be lower than 186 MHz). To
support DRA76's wide screen HDMI feature, we need to increase this
maximum rate.

This patch is a hack until we get a clear understanding from the HW
people on what are the limits. Testing shows that the PLL seems to work
fine even with ~240MHz clocks, and even the HDMI output at that clock is
stable enough for monitors to show a picture. This holds true for all
DRA7 and AM5 SoCs (and probably also for OMAP5).

However, the highest we can go without big refactoring to the clocking
code is 192MHz, as that is the DSS func clock we get from the PRCM. So,
increase the max HDMI pixel clock to 192MHz for now, to allow some more
2k+ modes to work.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>