phy: ti: phy-ti-am654: Fix boot lockup
At commit ec5f414b0cb ("phy: ti: phy-ti-am654: Fix compilation warning")
we experience lockup at boot. This is because pr_warn() accesses
hw->init->name and that seems to be an invalid pointer.

The real reason for that is that hw->init was being allocated on the stack and
it would get corrupted. Fix this by allocating memory for hw->init properly.

Thanks to Peter Ujfalusi for assisting with debug.

Fixes: ff518e356cd ("phy: ti: phy-ti-am654: Add support to select PLL mux and left/right output mux")
Reported-by: Carlos Hernandez <ceh@ti.com>
Tested-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>