phy: ti: phy-ti-am654: Fix "Failed to find the parent" warnings
serdes_am654_clk_mux_get_parent() is not doing the right thing
by returning a valid clock index "0" if a valid parent was not found.

.get_parent() must always return a valid parent. There is no room
for error there.

The reason for not getting a valid parent was that the reset configuration
of the mux is 0 and it is not always mapped as a valid setting
in the mux_table. Fix this by always setting a supported
configuration before registering the clock mux.

Increase the severity of the warning if we couldn't find a supported
parent in serdes_am654_clk_mux_get_parent().

Fixes: ff518e356cd ("phy: ti: phy-ti-am654: Add support to select PLL mux and left/right output mux")
Tested-by: Peter Ujfalusi <peter.ujfalusi@ti.com>
Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>