TEMP: samples/rpmsg: add compatible to support J7ES PDK firmware images
Add a specific compatible "ti.ipc4.ping-pong" to the rpmsg client device
list so that the rpmsg client sample device published and used by the J7ES
PDK IPC sample firmware images can properly probe the rpmsg_client_sample
driver and provide a means to unit-test the rpmsg communication with the
various remote processors on J7ES SoCs.

This is needed because the current firmwares do not have support for the
enhanced name service announcement mechanism, and are not using the
standard "rpmsg-client-sample" device.

Signed-off-by: Suman Anna <s-anna@ti.com>