dmaengine: ti: k3-udma: Use dma_pool for packet mode channel descriptors
If the channel is configured in packet mode the cppi5 descriptor sizes are
fixed (HPD, HBD). Instead of the expensive dma_zalloc_coherent use dma_pool
for the cppi5 descriptors.
The dma_pool will allocate memory on demand providing the channel enough
space to work without the need to allocate memory over time, which saves
time in prepare time.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>