remoteproc/k3-r5: add support for fixed reserved memory regions
The R5F remote processors do not have an MMU, and so require fixed
memory carveout regions matching the firmware image addresses. The
R5F remoteproc driver currently relies on a single memory region
provided to the device and uses it to dynamically allocate memory
for both the vrings/vring buffers as well as for the firmware image
segments. The latter is done through matching RSC_CARVEOUT entries.
This is not a scalable solution though as there is no way to ensure
the allocated memories will match the exact firmware image addresses.

Enhance the current R5F remoteproc driver to remove this limitation.
Support for this is added by mandating multiple memory regions to be
attached to the remoteproc device. The first memory region will be
used to serve as the DMA pool for all dynamic allocations like the
vrings and vring buffers. The remaining memory regions are mapped
into the kernel at device probe time, and are used to provide address
translations for firmware image segments without the need for any
RSC_CARVEOUT entries. Any firmware image using memory outside of
the supplied reserved memory carveout regions will be errored out.

NOTE:
The R5F remoteproc driver currently does not support the Region
Address Translator allowing the R5F cores to use a 32-bit processor
address for leveraging a physical address > 32-bit address. All
the reserved memory regions are therefore expected to be using
memory regions within the first 2 GB.

Signed-off-by: Suman Anna <s-anna@ti.com>