dmaengine: ti: k3-udma: Remove unnecessary dev_width <-> elsize conversions
There is no point of converting dma_slave_buswidth to elsize, then when
needed convert elsize to number of bytes.
The dma_slave_buswidth is telling the number of bytes.

Furthermore for Static TR we can save code to do the dma_slave_buswidth to
elsize conversion when configuring the static_tr.

Lastly we can remove the whole elsize related defines and struct.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>