drm/omap: omap_wb: remove unneeded vb2_dma_addr_plus_data_offset()
vb2_dma_addr_plus_data_offset() was originally created for the VPE
driver as a hack to handle NV12 frames and make them look like 2 plane
buffer by mis-using the .data_offset to point to the beginning of the UV
part of the image for the second plane. However .data_offset was meant
to be used by userspace using frame buffer which would contain an info
header where the actual video data starts. Sometime after kernel v3.14
the V4L2 framework started to check that the .data_offset could not be
larger than the .bytesused, meaning .data_offset could not indicate an
offset outside of the current buffer. The V4L2 framework nows issue WARN
if data_offset is anything but zero.

Since .data_offset is forced to zero by the framework then this change
does not change the code behavior.

Both VPE and OMAP WB-M2M driver were updated to properly deal NV12
format but the above helper function was not removed from the OMAP WB
M2M driver when we removed it from VPE. We are now correcting that
mistake.

Signed-off-by: Benoit Parrot <bparrot@ti.com>