remoteproc/k3-r5: cleanup k3_r5_cluster_rproc_exit() for split-mode
The function k3_r5_cluster_rproc_exit() performs complimentary operations
to those performed in the k3_r5_cluster_rproc_init() function and is
responsible for unregistering and freeing the remoteprocs. This function
currently unregisters the remoteprocs in the same order as done in the
init function. Reorder this unregistration so that Core1 is unregistered
before Core0 in Split-mode, the behavior in LockStep mode is unchanged
since this mode has only one remoteproc registered.

Signed-off-by: Suman Anna <s-anna@ti.com>