remoteproc/k3-r5: cleanup failure paths in k3_r5_cluster_rproc_init()
The k3_r5_cluster_rproc_init() function is not performing all the
required cleanup for failure paths if the MCU R5F cluster is configured
for Split-mode. Clean up the function so that all prior successful
operation on any of Core0 and/or Core1 are addressed upon any failure
on any core in both LockStep and Split modes. The changes also fixes
a minor cleanup bug introduced in commit 2419f21396b9 ("remoteproc/k3-r5:
fix the R5 boot state machine for multiple combinations").

Signed-off-by: Suman Anna <s-anna@ti.com>