remoteproc/k3-r5: fix missing put_device() in cluster OF init code
The of_find_device_by_node() function returns a platform device for the
passed in device node and takes a reference to the embedded struct device
on success. This needs to be dropped to balance the reference count once
the platform device usage is done with. This is currently missing on each
of the R5F core nodes in the k3_r5_cluster_of_init() function and results
in a memory leak. Fix this properly.

Fixes: 8cdfa3a5d4b8 ("remoteproc/k3-r5: add a remoteproc driver for MCU (dual-R5F) subsystem")
Signed-off-by: Suman Anna <s-anna@ti.com>