ti_config_fragments: rpmsg: Enable OMAP remoteproc support
Add support to build the OMAP remoteproc driver and the dependent
OMAP IOMMU driver. The dependent REMOTEPROC is already enabled as
built-in in the needed multi_v7_defconfig defconfig. The other
dependent OMAP Mailbox driver is automatically selected by the
OMAP remoteproc driver and is also already enabled in the current
config fragment. This driver is required for remote processor
messaging with the IPU and DSP remoteprocs on various OMAP4+ SoCs.

OMAP Mailbox and OMAP IOMMU are chosen to be built-in, while the
OMAP remoteproc driver is chosen to be built as a module. The IOMMU
debugfs configuration option is also also enabled by default.

Signed-off-by: Suman Anna <s-anna@ti.com>