drm/tidss: static partitionning of common areas
DSS7 has support for partitioning the common areas
among the several compute entities and they can be
independently driven from different cores.

This patch adds support for statically partitioning
the common areas using device-tree nodes.

If there is a DT node that specifies which common areas
to use and what functionality, then the information is
used to access / bypass the specified common area

Signed-off-by: Subhajit Paul <subhajit_paul@ti.com>