drm/tidss: fix handling of no-PM case
If the kernel is configured without PM, runtime PM calls will be no-ops.
However, the tidss driver relies on runtime PM resume and suspend to
function properly.

We have a special case inside #ifdefs to manually call the runtime
resume and suspend so that the driver initializes correctly.

However, the driver is checking against CONFIG_PM_SLEEP, which is not
correct. This patch fixes the check to be against CONFIG_PM.

Signed-off-by: Jyri Sarha <jsarha@ti.com>
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Reviewed-by: Benoit Parrot <bparrot@ti.com>