dmaengine: ti: k3-udma: Make slave TR mode operational with PDMAs
If we use TR mode for peripherals (serviced by PDMA) which is not a
streaming type (cyclic) then PDMA will not return all the data it has
received from the remote end causing the RX operation to stall.

However telling PDMA to close the packet and flush it's FIFO makes the TR
slave mode operational with the same limitation as with packet mode.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>