drm/tidss: dispc7: Add support for j721e DSS
This patch adds support for DSS on J721E, which is a SoC in Texas
Instruments Keystone 3 family.

J721E DSS is DSS7, similarly to AM6. However, the J721E DSS7 is an
extended version, and while the basic features are the same, there are
multiple differences. Some of the features/differences are:

- 4 videoports
- 4 planes (2 of the "lite", without scaling)
- Support up to 8k displays
- 2 DPI outputs
- DisplayPort and DSI support (implemented separately as DRM bridge
  drivers).
- No OLDI
- Support for DSS videport/plane sharing between different CPU cores

Signed-off-by: Jyri Sarha <jsarha@ti.com>
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Reviewed-by: Benoit Parrot <bparrot@ti.com>