arm64: dts: ti: k3-j721e-som-p0: Add HyperFlash node
J721e SoM as a 64MB HyperFlash on board. Add pinmux and DT node for the
same.
OSPI and HBMC are muxed within SoC's Flash Subsystem(FSS) hence only one
of them can be used at any time. But, J721e SoM board has both OSPI
flash and HyperFlash on it. There is an onboard switch that selects
b/w OSPI flash and HyperFlash which can be read to know which flash is
active. By default, this mux is set to OSPI flash. Hence, keep HBMC
node disabled in DT. Bootloader will enable appropriate node if it
detects that onboard mux is set to HyperFlash instead of OSPI flash.

Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>