drm/bridge: cdns-mhdp: Simplify DP output color format hacks
There is no use to look what is in connector's display_info, since the
driver currently only supports DRM_COLOR_FORMAT_RGB444 and 8 bit per
color. The patch adds a warning print if this selected format is not
supported by the connected display.

Cdns-mhdp can not do color format conversions. So the proper
implementation for choosing the video sink color format should be made
in the DRM display driver and be communicated down to the cdns-mhdp
bridge.

Signed-off-by: Jyri Sarha <jsarha@ti.com>
Reviewed-by: Tomi Valkeinen <tomi.valkeinen@ti.com>