ti_config_fragments/connectivity.cfg: enable DaVinci NAND explicitly
DaVinci NAND support is disabled on Keystone2 platforms which
need it.

This happened because multi_v7_defconfig does not enable
CONFIG_MEMORY anymore because its selected by
CONFIG_ARCH_OMAP2PLUS (which is enabled). But we disable
CONFIG_ARCH_OMAP2PLUS when building for keystone2.

fix this by enabling the dependencies needed for NAND
support on keystone2 explicitly.

Signed-off-by: Sekhar Nori <nsekhar@ti.com>