arm64: dts: ti: k3-j721e-main: Add DMA entries for main_uart4
Add DMA entries for main_uart4. Since there aren't enough DMA channels
allocated to Linux to support DMA on all UART instance, add entries only
for main_uart4 to serve as an example. Reason for choosing main_uart4 is
that it is available for testing on an header on GESI card.

Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>