remoteproc/k3-dsp: Fix successive boots of C66x DSPs in remoteproc mode
The C66x DSP remote processors on J721E SoCs have both a local reset
and a global module reset. The local reset is deasserted on SoC power
on reset, and so did not require any additional logic to achieve the
first remoteproc-mode boot. The local reset does control the CPU reset,
and so the local reset needs to be asserted and deasserted for properly
shutting down and re-booting the C66x DSP in remoteproc mode using either
the same firmware or a different firmware.

The core logic for this is already in place but disabled under a
pre-processor macro. Clean up the macro and fix the C66x successive
boot.

NOTE:
 - The C71x DSP is not affected by this change.
 - The current code still does not support loading into DSP internal
   L1 or L2 RAMs, this logic will require further refactoring of code.

Signed-off-by: Suman Anna <s-anna@ti.com>