ASoC: ti: Add custom machine driver for j721e EVM (CPB and IVI)
The audio support on the board is using pcm3168a codec connected to McASP10
serializers in parallel setup.
The pcm3168a SCKI clock is coming via the j721e AUDIO_REFCLK2 pin.
In order to support 48KHz and 44.1KHz family of sampling rates the parent clock
for AUDIO_REFCLK2 needs to be changed between PLL4 (for 48KHz) and PLL15 (for
44.1KHz). The same PLLs are used for McASP10's AUXCLK clock via different
HSDIVIDER.

Generic card can not be used for the board as we need to switch between
clock paths for different sampling rate families and also need to change
the slot_width between 16 and 24 bit audio.

The audio support on the Infotainment Expansion Board consists of McASP0
connected to two pcm3168a codecs with dedicated set of serializers to each.
The SCKI for pcm3168a is sourced from j721e AUDIO_REFCLK0 pin.
It is extending the audio support on the CPB.

To handle the two codecs on the expansion board we will have separate PCMs
for each:
hw:0,0 - cpb playback
hw:0,1 - cpb capture
hw:0,2 - ivi cocec-a playback
hw:0,3 - ivi cocec-a capture
hw:0,4 - ivi cocec-b playback
hw:0,5 - ivi cocec-b capture

Due to the fact that the same PLL4/15 is used by both domains (CPB/IVI)
there are restriction on sample rates and also the two codec on the IVI
can not be used at the same time.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>