irqchip/irq-pruss-intc: Add support for ICSSG INTC on K3 J721E SoCs
The K3 J721E SoCs have a revised version of the ICSSG IP present on K3
AM65x SoCs. The J721E ICSSG INTC is identical to that of the AM65x ICSSG
INTC despite the presence of two new additional auxiliary Tx_PRU cores.
The Tx_PRU cores reuse the same host interrupts as the regular PRU cores,
so the recommended approach is to use the interrupt mapping configuration
through the corresponding PRU core.

Enhance the PRUSS INTC driver to add support for this ICSSG INTC instance.
This support is adding using a new SoC-specific compatible and reusing
the AM65x support code.

Signed-off-by: Suman Anna <s-anna@ti.com>
Reviewed-by: Roger Quadros <rogerq@ti.com>