gpu: Add carveouts for virtual GPU driver
The GPU can be used by multiple virtual machines in a hypervisor
environment. Part of the setup requires that the GPU has contiguous
memory allocated for the host and guest firmwares, where each region is
the same size.

Add the necessary carveouts to have the GPU firmware memory regions
reserved only if using the jailhouse setup.

Requires GPU to be enabled on the host/root-cell, so re-enable it in
the jailhouse overlay.

Acked-by: Nikhil Devshatwar <nikhil.nd@ti.com>
Signed-off-by: Erick Narvaez <e-narvaez@ti.com>
Signed-off-by: Gowtham Tammana <g-tammana@ti.com>